+++ src/plugins/explore/public/components/query_panel/components/footer/run_query.test.tsx
@@ -1,3 +1,8 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
 import React from 'react';
 import { render, fireEvent, screen } from '@testing-library/react';
 import { RunQueryButton } from './run_query';
