+++ src/plugins/explore/public/components/query_panel/components/footer/show_input_type.test.tsx
@@ -1,3 +1,8 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 import { ShowInputType } from './show_input_type';
