+++ src/plugins/explore/public/components/query_panel/components/footer/switch_lanugage.test.tsx
@@ -1,3 +1,8 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
 import React from 'react';
 import { render, fireEvent, screen } from '@testing-library/react';
 import { SwitchLanguage } from './switch_language';
